Hide analyses on closed pull request (#33)
For closed pull requests (aka merged or rejected), it may not be possible to show anaysis results. This may be, because either the branch and/or the revision no longer exists. Even if we can find analysis results, they may be out of date or irrelevant. So they should not be shown in closed pull requests any longer.

Further more this prevents a "not found" error, when no changeset can be found for a pull request.