Fix NPE for unknown revision (#43)
Fixes a NullPointerException that is thrown when the revision in a CI status PUT request is given completely, but does not exist. This exception will not be thrown in Git repositories for shortened revisions, because in this case another NullPointerException is catched in the Git plugin and will be converted into a NotFoundException there.