Add status flag for replacement
This fixes an issue with Jenkins multibranch pipelines.
If such a pipeline fails for a branch for whatever
reason (eg. due to a flappy test), this result is stored
for the revision. If later on a pull request is created
for this branch, this failed build is considered as a
failed status for the pull request, even though the
pull request build itself might be successful.
    
To fix this, a new field 'replaces' is added to the
 status. This can be set to a name of another
potentially existing status, that will than be
ignored for the pull request (if the type is the same).

Committed-by: Konstantin Schaper <konstantin.schaper@cloudogu.com>
Co-authored-by: René Pfeuffer <rene.pfeuffer@cloudogu.com>