Fix error when creating CI status with spaces in the name
When a CI status where the name (or type as it is) contains spaces, in
the URL these spaces would be encoded to '+' characters. These pluses
would not be decoded to spaces again in the REST resource and therefor
would be treated as different from the name (or type) specified in the
dto.

With this fix, we simply replace all plus characters with spaces in the
check and therefore ignore differences in these. Other characters
like parens or such would be encoded too, but will be decoded
correctly again and therefor do not have to be treated specially.